Further minor compile error fixes for 32 bit builds

+++ containers.h
@@ -166,7 +166,7 @@ struct TokenList : public Container<TokenList_t>, public Iterable {
     if (list().size() <= idx) {
       throw std::out_of_range("List index out of range!");
     }
-    return list()[idx];
+    return list()[(size_t)idx];
   }
 
   void push(packToken val) const { list().push_back(val); }
